Ensure connection always get properly initialized.
Previously it was possible that `getcursor` might get a connection
which was not correctly initialized. This occurred when the last
connection in the pool got closed before calling `getcursor`. As in this
case the code guarded by `init` in `getconn` was not called.

Add support for PostgreSQL 13. (It might raise InterfaceError exceptions)